Add upper constraints
This is used by the OSA deployment. This file does not really belong
into the Octavia repo, but as we use a newer OpenStack Release for
Octavia than for other services and we have some modifications to the
constraints, we need to keep this file somewhere. Copying here is easier
than creating a fork of the OpenStack requirements repository for this.

This file can be removed, once the rest of the OpenStack services are
upgraded to Yoga or there are no more modifications from upstream.

Change-Id: Ie394c2177e18c60576376f967ea1f201d62b799f