Lock SELECTed DB rows when updating
The use of with_for_update() when querying the DB for an update was
only applied for some updates and not for others. This commit adds
populate_existing() and with_for_update() to all places found where a
query result is later used in an update.

This fixes a concurrency issue when deleting the same pool member with
2 requests at the same time. It probably also fixes other similar
concurrency issues.

Change-Id: I39f0a89dbb3df468c561abea750fe8b5bc079d87
(cherry picked from commit 3b64fdad2b71e1833c518ed266731791189ce811)